Issue #181 - ns metadata hashmap with quoted value
This commit fixes an issue with ns metadata hashmap when the value is a
quoted map.

After collecting the metadata value, we need to skip until the end of
the collected values. If the node is a wrap node, the lastChildNode.id
is the wrap node body. We need to skip more than that to the end of the
wrap node.